+++ magento/order_helpers.py
@@ -1,6 +1,4 @@
-from typing import Any, Dict
-
-from magento.types import Order
+from magento.types import Order, ShippingAddress
 
 __all__ = (
     'is_order_on_hold',
@@ -17,16 +15,16 @@ def is_order_on_hold(order: Order) -> bool:
 def is_order_cash_on_delivery(order: Order) -> bool:
     """Test if an order is paid with 'cash-on-delivery'."""
     # From Magento\OfflinePayments\Model\Cashondelivery::PAYMENT_METHOD_CASHONDELIVERY_CODE
-    # For some reason Mypy is too dumb to see this as boolean
+    # For some reason Mypy is too dumb to see this is a boolean
     b: bool = order["payment"]["method"] == 'cashondelivery'
     return b
 
 
-def get_order_shipping_address(order: Order) -> Dict[str, Any]:
+def get_order_shipping_address(order: Order) -> ShippingAddress:
     """Return the first shipping address of an order.
 
     Note the returned dict is a reference, so if you modify it, it modifies the order.
     Make a copy if you want to modify the address without affecting the order.
     """
-    address: Dict[str, Any] = order["extension_attributes"]["shipping_assignments"][0]["shipping"]["address"]
+    address: ShippingAddress = order["extension_attributes"]["shipping_assignments"][0]["shipping"]["address"]
     return address